feat(core): migrate parameters for the host call fetch (#84)
* feat(specjam): approve scale for history and preimages

* chore(core): update the types of parameters

* feat(pvm): implement the fetch call with kind 0 supported

* chore(pvm): use vlen instead of calling len() each of times in general calls

* feat(pvm): pass operands to general context

* fix(memory): return status of previous storage value

* chore(pvm): handle returns of storage write

* feat(runtime): order the storage ops

* chore(core): update D for tiny testnet

* chore(jade): remove jade from workspace

* feat(core): introduce static paramaters

+++ crates/core/src/block/mod.rs
@@ -58,7 +58,7 @@ impl Block {
 
 /// Returns the current timeslot
 pub fn timeslot() -> TimeSlot {
-    now() / crate::SLOT_PERIOD
+    now() / (crate::SLOT_PERIOD as u32)
 }
 
 /// Returns the current time in seconds
